Update GameServer.State.cs
Build.bat appears to be using a compiler that's too out of date to parse the "?." syntax, so it was failing to build with the changes that had been made to the card rarity section of GameServer.State.cs. In lieu of altering the build.bat file I've simply replaced the "?." code in this file with equivalent null checks.